requirements.txt
selenium==4.23.1
pytest==8.3.2
pyautogui==0.9.54
requests==2.32.3
allure-pytest==2.13.5
pytest-xdist==3.6.1